:sparkles: :zap: deleting files

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../utils/file');
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -167,6 +169,7 @@ exports.postEditProduct = (req, res, next) => {
 			}
 			product.title = updatedTitle;
 			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
 				product.imageUrl = image.path;
 			}
 			product.price = updatedPrice;
@@ -182,7 +185,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const productId = req.body.productId;
-	Product.deleteOne({ _id: productId, userId: req.user._id })
+	Product.findById(productId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error('Product not found.'));
+			}
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: productId, userId: req.user._id });
+		})
 		.then(() => res.redirect('/admin/products'))
 		.catch((err) => {
 			const error = new Error(err);
